refactor(contracts): extract RPC schemas
