+++ packages/contracts/src/context.ts
@@ -1,6 +1,7 @@
 import { z } from "zod"
 import { ChatMessageSchema } from "./chat"
 
+/** Text already extracted from a user file for context construction. */
 export const ContextFileInputSchema = z.object({
   text: z.string(),
   metadata: z.object({
@@ -11,6 +12,11 @@ export const ContextFileInputSchema = z.object({
 
 export type ContextFileInput = z.infer<typeof ContextFileInputSchema>
 
+/**
+ * Persistable context-build command. Callback, toast, browser, and other
+ * application-only state is deliberately excluded so a suspended background
+ * runtime can resume the same turn from this value alone.
+ */
 export const DurableContextOptionsSchema = z.object({
   rawInput: z.string(),
   files: z.array(ContextFileInputSchema).optional(),
@@ -39,10 +45,12 @@ export const DurableContextOptionsSchema = z.object({
   customModel: z.string().optional()
 })
 
+/** Persisted context shape before application message normalization. */
 export type DurableContextOptionsParsed = z.infer<
   typeof DurableContextOptionsSchema
 >
 
+/** Parse a persisted context command without application environment access. */
 export const parseDurableContextOptions = (
   value: unknown
 ): DurableContextOptionsParsed => DurableContextOptionsSchema.parse(value)
