+++ packages/contracts/src/index.ts
@@ -1,6 +1,8 @@
 export * from "./app-failure"
 export * from "./chat"
 export * from "./context"
+export * from "./model-rpc"
+export * from "./provider-rpc"
 export * from "./rpc"
 export * from "./streams"
 export * from "./turns"
