+++ src/features/model/hooks/use-provider-models.ts
@@ -1,3 +1,7 @@
+import {
+  MODEL_DISCOVERY_FAILURE,
+  type ProvidersListModelsResult
+} from "@ollama-client/contracts/provider-rpc"
 import { useStorage } from "@plasmohq/storage/hook"
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query"
 import { useCallback, useEffect } from "react"
@@ -24,10 +28,6 @@ import {
 } from "@/lib/providers/types"
 import { queryKeys } from "@/lib/query-keys"
 import { extensionRpcClient } from "@/protocol/extension-client"
-import {
-  MODEL_DISCOVERY_FAILURE,
-  type ProvidersListModelsResult
-} from "@/protocol/provider-rpc"
 import { RpcMethod } from "@/protocol/rpc"
 import type { SelectedModelRef } from "@/types"
 import { isEmbeddingModel } from "../lib/model-utils"
