+++ src/lib/constants/sqlite.ts
@@ -1,5 +1,4 @@
-// SQLite Database Storage Constants
-// Legacy database name retained for compatibility with existing installs.
+/** Legacy database name retained for compatibility with existing profiles. */
 export const SQLITE_DB_NAME = "ollama-sqlite-db"
 export const SQLITE_DB_STORE = "sqlite-store"
 export const SQLITE_DB_KEY = "database"
