+++ src/sidepanel/components/first-run-permissions-dialog.tsx
@@ -44,10 +44,6 @@ import {
 } from "@/lib/onboarding/state"
 import { saveSelectedModelRef } from "@/lib/providers/selected-model"
 import { extensionRpcClient } from "@/protocol/extension-client"
-import type {
-  ProvidersListModelsResult,
-  PublicProviderConfig
-} from "@/protocol/provider-rpc"
 import { RpcMethod } from "@/protocol/rpc"
 
 const STAGES: OnboardingStage[] = [
@@ -462,3 +458,8 @@ export const FirstRunPermissionsDialog = () => {
     </>
   )
 }
+
+import type {
+  ProvidersListModelsResult,
+  PublicProviderConfig
+} from "@ollama-client/contracts/provider-rpc"
